existance check

+++ elements/keepalived-status-check/pre-install.d/status_check.py
@@ -28,7 +28,8 @@ def getNeutronClient():
 def is_keepalived_master():
     retry_cnt = 0
     data_path = "/tmp/keepalived.data"
-    os.remove(data_path)
+    if os.path.isfile(data_path):
+      os.remove(data_path)
     pid = sorted(list(map(int,check_output(["pidof", "keepalived"]).split())))[0]
     os.kill(int(pid), signal.SIGUSR1)
     while not os.path.exists(data_path):
